ReType_BatchFontInfo-1.0.py
This script fill the most important fields in a font family at once in Fontlab. The program follows Karsten Lücke's recommendations about Font Metrics and Font Naming in his well known PDFs. I will try to improve the interface and the script in the future, but for the moment, this is all I can do. I'm sure several of you will point better methods or syntax :) so thanks a lot in advance. 
Thanks to Tal Leming and Paul van der Laan for the help. 

Usage: 
1- Open your fonts
2- Set the "Weight" in each font info and click the "Italic" checkbox (if the font is italic). If you want to use different font weight names to the ones available in the Fontlab menu, the script must be edited. The script will uncheck the boxes later (it does not make linking families)
3- Run the script, complete the required information and press "OK".
4- Go and check the font info in each font to see the changes.
Warning: Robofab and Dialog Kit must be installed.